MOBILE-2951 web-services: make it easier to debug web service calls
We add a redundant &wsfunction=web_service_method_name to the URL
for web service requests, so that it is easier to find the call you
are interested in when debugging using brower tools.

Similarly, we add info= to calls to lib/ajax/service.php, matching what
is done in Moodle web application JavaScript.