MOBILE-3180 Site plugins: Plugin pages offline can break navigation
For pages served via a site plugin, if the page has never been viewed
(there is no cached version) then if you try to view the page when
offline, the navigation broke because page loading never finished.

This change makes it finish page loading (with a blank page) if it
shows the error about failing to load it.