MOBILE-3876 test: Fix compatibility checks
Using the new minification strategy, it finds the unsupported ES2018 RegExp on src/core/services/utils/text.ts:333. It's actually fine because it is written within a try/catch, so we can ignore it. But it should have been detecting it all along, so it was a false positive before.